Replace types in MemoryManager modules
