+++ lamagraph-core/src/Core/Reducer.hs
@@ -6,25 +6,12 @@ module Core.Reducer where
 
 import Clash.Prelude
 import Control.Lens
+import Core.MemoryManager.MemoryManager
+import Core.MemoryManager.NodeChanges
 import Core.Node
 
 type NumOfNodesToStore = Unsigned 3
 type NumOfEdgesToStore = Unsigned 3
-
-data Edge (portsNumber :: Nat) = Edge
-  { _leftEnd :: Connection portsNumber
-  , _rightEnd :: Connection portsNumber
-  }
-  deriving (Generic, NFDataX, Show, Eq)
-
-$(makeLenses ''Edge)
-
-data ActivePair (portsNumber :: Nat) = ActivePair
-  { _leftNode :: LoadedNode portsNumber
-  , _rightNode :: LoadedNode portsNumber
-  }
-  deriving (Show, Eq, Generic, NFDataX, Bundle)
-$(makeLenses ''ActivePair)
 
 -- | Result of abstract reduction rule
 data ReduceRuleResult (nodesNumber :: Nat) (edgesNumber :: Nat) (portsNumber :: Nat) = ReduceRuleResult
@@ -33,17 +20,6 @@ data ReduceRuleResult (nodesNumber :: Nat) (edgesNumber :: Nat) (portsNumber :: 
   }
 
 $(makeLenses ''ReduceRuleResult)
-
-data Delta (nodesNumber :: Nat) (edgesNumber :: Nat) (portsNumber :: Nat) = Delta
-  { _newNodes :: Vec nodesNumber (Maybe (LoadedNode portsNumber))
-  , _newEdges :: Vec edgesNumber (Maybe (Edge portsNumber))
-  , _activePair :: ActivePair portsNumber
-  }
-  deriving (Show, Eq, Generic, NFDataX)
-
-$(makeLenses ''Delta)
-
-type Interface externalNodesNumber = Vec externalNodesNumber (Maybe AddressNumber)
 
 toDelta ::
   (KnownNat portsNumber, KnownNat edgesNumber, KnownNat nodesNumber) =>
